Create generic RAS interface, Update RAS classes and rename variables and methods
