+++ src/main/java/com/bayou/ras/impl/UnverifiedUserResourceAccessor.java
@@ -1,6 +1,7 @@
-package com.bayou.ras;
+package com.bayou.ras.impl;
 
 import com.bayou.domains.UnverifiedUser;
+import com.bayou.ras.IResourceAccessor;
 import com.bayou.repository.IUnverifiedUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
